+++ frontend/src/layouts/DesignLayout/index.tsx
@@ -230,7 +230,7 @@ const DesignLayout: React.FC<DesignLayoutLayoutProps> = props => {
               }}
             >
               <div>{project.projectName}</div>
-              <div>© 2024 Made with 零代科技</div>
+              <div>© 2026 ERD Online · MIT</div>
               <div>ERD Online</div>
             </div>
           );