+++ frontend/src/pages/account/settings/index.tsx
@@ -148,7 +148,7 @@ const Settings: React.FC = () => {
                 paddingBlockStart: 12,
               }}
             >
-              <div>© 2024 Made with 零代科技</div>
+              <div>© 2026 ERD Online · MIT</div>
               <div>ERD Online</div>
             </div>
           );